StrongFieldDynamics

Documentation for StrongFieldDynamics.

A Julia package for studying intense-laser atom interactions. Julia is fast.

StrongFieldDynamics.ContinuumSolution

StrongFieldDynamics.ClebschGordan

StrongFieldDynamics.F1_integral

StrongFieldDynamics.F2_integral

StrongFieldDynamics.T0

StrongFieldDynamics.Ylm

StrongFieldDynamics.atomic_electron

StrongFieldDynamics.bessel_electron

StrongFieldDynamics.distorted_electron

StrongFieldDynamics.inner_product

StrongFieldDynamics.probability

StrongFieldDynamics.reduced_matrix_element

StrongFieldDynamics.sin2Sv

StrongFieldDynamics.ContinuumSolution — Type

ContinuumSolution

Bessel -> ... Coulomb -> ... Distorted -> ...

StrongFieldDynamics.ClebschGordan — Method

ClebschGordan(ja, ma, jb, mb, Jab, Mab)

Computes the Clebsch-Gordan coefficients

StrongFieldDynamics.F1_integral — Method

The pulse shape integral

StrongFieldDynamics.F2_integral — Method

The pulse shape integral

StrongFieldDynamics.T0 — Method

T0(pulse::Pulse, a_electron::AtomicElectron, p_electron::ContinuumElectron, mj::Rational{Int64}, msp::Rational{Int64}, θ::Float64, ϕ::Float64)

Computes the direct scattering amplitude

StrongFieldDynamics.Ylm — Method

Ylm(l, m, θ, ϕ)

Computes the Spherical hamonics using the Associated Legender Polynomials as Ylm(l, m, θ, ϕ) = Plm(l, m, cos(θ)) * exp(immϕ)

StrongFieldDynamics.atomic_electron — Method

atomic_electron(n::Int64, l::Int64)

Computes the atomic electron wavefunction for the corresponding n & l Principal and Orbital quantum numbers, repectively.

StrongFieldDynamics.bessel_electron — Method

bessel_electron(ε::Flost64, l::Int64, r::Vector{Float64})

Computes the radial part of the continuum (free) electron with Bessel function P (r) = r jₗ(pr) Returns the radial wavefunction of type Vector{Float64}

StrongFieldDynamics.distorted_electron — Method

distorted_electron(ε::Float64, l::Int64, r::Vector{Float64}, pot::Vector{Float64})

Computes the distorted wave function of continuum photo electron with Bessel function Returns the radial wavefunction and phase-shift of type (Vector{Float64}, Float64)

StrongFieldDynamics.inner_product — Method

inner_product(p_partialwave::PartialWave, a_electron::AtomicElectron, r::Vector{Float64})

Calculates the inner product of continuum and bound (atomic) electron radial wavefunction

StrongFieldDynamics.probability — Method

probality(pulse::Pulse, a_electron::AtomicElectron, p_electron::ContinuumElectron)

Calculates the ionization probability

StrongFieldDynamics.reduced_matrix_element — Method

reduced_matrix_element(εp, lp, jp, n, l, j)

Computes the reduced matrix element <εp lp jp || p || n l j>

StrongFieldDynamics.sin2Sv — Method

sin2Sv(t::Float64, θ::Float64, pulse::Pulse, a_electron::AtomicElectron, p_electron::ContinuumElectron)

Defines a sin² envelope.